requirements.txt
azure-cognitiveservices-vision-computervision
beautifulsoup4
dj-database-url
Django
emoji
psycopg[binary]
python-twitter-v2
requests
streamlit
urlextract